auth: refactor code related to the authorization URL
