+++ googleapis_auth/lib/src/oauth2_flows/auth_code.dart
@@ -3,7 +3,11 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'dart:async';
+import 'dart:convert';
+import 'dart:math';
+import 'dart:typed_data';
 
+import 'package:crypto/crypto.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/http.dart';
 
@@ -12,6 +16,71 @@ import '../client_id.dart';
 import '../exceptions.dart';
 import '../known_uris.dart';
 import '../utils.dart';
+
+Uri createAuthenticationUri({
+  required String redirectUri,
+  required String clientId,
+  required Iterable<String> scopes,
+  required String codeVerifier,
+  String? hostedDomain,
+  String? state,
+  bool offline = false,
+}) {
+  final queryValues = {
+    'client_id': clientId,
+    'response_type': 'code',
+    'redirect_uri': redirectUri,
+    'scope': scopes.join(' '),
+    'code_challenge': _codeVerifierShaEncode(codeVerifier),
+    'code_challenge_method': 'S256',
+    if (offline) 'access_type': 'offline',
+    if (hostedDomain != null) 'hd': hostedDomain,
+    if (state != null) 'state': state,
+  };
+  return googleOauth2AuthorizationEndpoint.replace(
+    queryParameters: queryValues,
+  );
+}
+
+/// https://developers.google.com/identity/protocols/oauth2/native-app#create-code-challenge
+/// Between 43 and 128 = 86
+String createCodeVerifier() {
+  final rnd = Random.secure();
+
+  return List.generate(86, (index) => _safe[rnd.nextInt(_safe.length)]).join();
+}
+
+/// See https://developers.google.com/identity/protocols/oauth2/openid-connect#createxsrftoken
+String randomState() {
+  final rnd = Random.secure();
+
+  final list = Uint32List(6);
+  for (var i = 0; i < list.length; i++) {
+    list[i] = rnd.nextInt(1 << 32);
+  }
+
+  final value = base64UrlEncode(Uint8List.view(list.buffer));
+  return _stripBase64Equals(value);
+}
+
+// https://datatracker.ietf.org/doc/html/rfc3986#section-2.3
+const _safe = '0123456789-._~'
+    'abcdefghijklmnopqrstuvwxyz'
+    'ABCDEFGHIJKLMNOPQRSTUVWXYZ';
+
+String _codeVerifierShaEncode(String codeVerifier) {
+  final asciiBytes = ascii.encode(codeVerifier);
+  final sha26Bytes = sha256.convert(asciiBytes).bytes;
+  final output = base64UrlEncode(sha26Bytes);
+  return _stripBase64Equals(output);
+}
+
+String _stripBase64Equals(String value) {
+  while (value.endsWith('=')) {
+    value = value.substring(0, value.length - 1);
+  }
+  return value;
+}
 
 // A successful response from the server will give an HTTP response status
 // 200 and a body of the following type:
